RAG_demo/requirements.txt
langchain==0.2.16
langchain-community==0.2.16
langchain-core==0.2.38
langchain-openai==0.1.20
langchain-text-splitters==0.2.4
faiss-cpu==1.7.4
PyPDF2==3.0.1
dashscope==1.14.0
numpy==1.24.0
typing-extensions==4.12.2
ipykernel==7.2.0
chromadb==0.5.13
json5==0.9.25
PyMuPDF==1.24.0
python-docx==1.1.2
docx2txt
matplotlib==3.8.3
pandas==2.1.4
seaborn==0.13.2
qwen-agent[rag,code_interpreter,gui]==0.0.9
